first mod Storybook
- checking in my Storybook mod and a tool I made for cleaning up the
BackstoryDef.xml

Basically I keep all my stuff in BackstoryDef.txt and About.txt and
those files are used to generate the .xml files when you run the Tool
for Storybook. Mainly I do this to keep About.xml updated with the
lastest backstories included as well as removing the 'author' tag I
added to backstories to keep track of who created it, as I'm using other
people's backstories as well.